Basic test execution
Test cases are specified in the YAML config file. Each should have the
following structure:

- name: <test name>
  score: <test score>
  args: <CLI arguments to run with>
  stdin: <string passed to stdin>
  stdout: <expected stdout>

Each test case is executed for each student solution. The solution
binary is run with given arguments and given stdin and its stdout is
checked to match the expected one. If the output matches, the given
score is added to the total solution score.