Introduce integration testing
Integration tests are located in the tests/ directory.

The tests have the form of testing projects, each project containing a
number of solutions with expected scores. Each solution is evaluated and
the obtained score is compared with the expected one.

The expected testing project directory structure is:
tests/projects/<project-name>/
  <solution-1>
  <solution-2>
  ...
  config.yaml
  expected-results

Where 'expected-results' has the form:
<solution-1>: <expected-score>
<solution-2>: <expected-score>
...

Tests for a single project are generated automatically using a
procedural macro. The macro is defined in a separate crate
(tests/test_generator).

When creating a new testing project, call a macro:
  generate_tests!("<project-name>");
from test_defs.rs.
Project name must be a valid Rust identifier.