Integration tests: better test generation
Previously, the file with expected scores for integration tests was
parsed during test generation (i.e. during compilation). This is not a
good approach since the tests are not re-generated when the file
changes.

This moves parsing of the scores into the individual tests, hence when
the expected score changes, the test is properly updated.

Note that the set of testing project solutions is still parsed during
test generation, therefore the tests must be force re-generated if a new
solution is added.