Allow specifying multiple test cases in one test
This is useful if some score should be awarded only if a (sub)set of
tests pass.

The config syntax is:
tests:
  - ...
    test-cases:
      - args: ...
        stdin: ...
        stdout: ...
      - args: ...
        ...
    require: <any|all>

The 'require' field specifies how many cases must pass to award the
points for the given test.

The original syntax where args, stdin, and stdout are given directly in
the test is still supported (one test case is created).